+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/BUILD.bazel
@@ -38,6 +38,7 @@ iree_compiler_cc_library(
         "Passes.cpp",
         "SplitReduction.cpp",
         "TileAndDecomposeAttention.cpp",
+        "TilingInterfaceImpl.cpp",
     ],
     hdrs = [
         "Passes.h",
